cgroup: trying to make rkt work on CentOS
We unmount /sys in rkt's mount namespace, so we can then simply create
/sys/fs/cgroup and mount the cgroup hierarchy there. This sounds good
but other problems appear.

* unmounting /sys works on CentOS 6 but on my Arch host, it returns
"device or resource busy". We'd have to unmount it recursively.

* nspawn uses `prctl(PR_SET_CHILD_SUBREAPER, 1)` which was introduced in Linux
3.4

* nspawn uses `newfstatat` with the flag `AT_EMPTY_PATH` which was introduced
in Linux 2.6.39

At that point I stopped trying. The testing process is very tedious and I'm
sure there're more things used in systemd that are not present in CentOS 6's
kernel (2.6.32).