rkt, build, docs: Rewrite stage1 image handling
It replaces the --stage1-image flag with five mutually-exclusive
flags: --stage1-url, --stage1-path, --stage1-name, --stage1-hash and
--stage1-from-dir. The flags tell rkt to use stage1 image,
respectively, from a URL, from a path, in the store under the name, in
the store with the hash, from a file in the default stage1 images
directory.

In case none of the above flags were provided, rkt takes stage1 and
paths configuration and build-time defaults to decide for which name
and version of the stage1 image it should search for in the image
store and, if that fails, from which location it should fetch the
image.

This introduces a new configure flag for defining the path to the
default stage1 images directory.