(change the path to rkt binary)tests: change the path to rkt binary for functional tests
we put a symlink to rkt binary in a directory used by functional
tests. in the same directory I also put a symlink to the default
stage1 image, so the tests should work as they did so far.

make clean considerations - need to create symlinks (valid or
dangling) to all possible flavors, so make clean could work even after
reconfiguration.