config: Add a new directory for stage0 configuration
This adds a new directory for stage0 configuration (for example
/etc/rkt/cfg/{paths.d,stage1.d,...}). This effectively deprecates the
old location, which means that new configuration kinds should be added
only to the new directory and the old directory should be kept frozen.

Currently, it will first try to get configuration from the new
directory and if there was none, rkt will fall back to the old
directory.