scripts: Minor fixes and a cleanup
Skip manifest.d directories in version bumping, because some parts of
shared library versions are matching the old rkt version and are
wrongly replaced with the new rkt version.

Add scripts/build-rir.sh and scripts/acbuild-rkt-builder.sh to version
bumping alongside the configure.ac file, because they are used for
building the rkt release.

Also, factor out the function that does the replacing.