functional tests: Split GPG key review to a separate function
Asking the user whether the user really trusts the key can happen in
two situation - when running rkt trust or when running rkt fetch with
--trust-keys-from-https set to false. Split out the function replying
to the question from the function calling rkt trust. The new function
will be used when testing various trust scenarios during rkt fetch.

Also, changed the function a bit to take a boolean parameter telling
whether to answer yes or no to the question.