functional tests: Add an exhaustive test for trusting keys
This basically tests all the combinations of signatures (missing,
signed wrong image with good key, signed wrong image with a wrong key,
good), keys (both to be trusted and on the server; missing, wrong or
good), trusting keys from https (rejecting the review, accepting the
review and actually trusting them).

It doesn't test the --signature parameter (which tells rkt to use some
local signature instead of the remote one), it probably could be done
in a separate test, to avoid complicating the test even more.

It is using stub stage1 image to get the 4-5x speedup, so even if
there are 108 combinations, it still should take less than a minute.