networking: Split some types into a separate package, cleanup
We will use ActiveNet and NetConf in yet another subpackage of
networking, which in turn will be used by networking. So to avoid
circular imports, move those types into the types subpackage.

At the same time, make all the code to use pointers to ActiveNet
instead of copying instances of ActiveNet everytime we are passing
them to some function or iterate the ActiveNet array.