networking: Port to pkg/config
This moves the configuration loading to config subpackage of the
networking package. Some hacks had to be added to preserve the
semantics of the previous configuration loading.

We will be deprecating it in favour of rkt-specific JSON config in a
separate, stage1-specific directory in a later commit.

Added some unit tests to make sure that the porting did not break
obvious things.