networking: Add new config type
This adds a support for rkt-specific configuration with "network" kind
and "v1" version. It allows overriding a priority, without the need
for specifying all the CNI config again.

The old configuration in /etc/rkt/net.d/42-foo.conf with contents as
follows:

```
{
  "name": "foo",
  "type": "bar",
  ...
}
```

could be translated to a new format, by putting an
/etc/rkt/stage1-cfg/net.d/foo.json file with the following contents:

```
{
  "rktKind": "network",
  "rktVersion": "v1",
  "name": "foo",
  "priority": 42,
  "cniConf": {
    "type": "bar",
    ...
  }
}
```