networking: Fix passing paths to CNI plugin
When tearing down networks during gc, the CNI_PATH variable sent with
the DEL command to the plugin is different than the one sent to the
plugin with the ADD command. Namely, local config directory is not
set, so we were basically sending "net.d" relative path to the plugin.

To fix, store the paths in the netinfo file and use them when if they
are available.