functional tests: Add a stub cni plugin
This plugin does not set any network, but rather creates a file with
some text as specified in its configuration. The config is basically:

```
{
  "rktKind": "network",
  "rktVersion": "v1",
  "name": "foo",
  "priority": 13,
  "cniConf": {
    "type": "stub-cni-plugin",
    "path": "/tmp/some-tmp-file",
    "text": "some text to be put in the file"
  }
```

This will make the plugin to write "some text to be put in the file"
to the "tmp/some-tmp-file" in the ADD command and delete this file in
the DEL command. This, in cooperation with inspect's "--read-stdin"
flag, will be useful to test custom search directories for network
plugins.