functional tests: Add a third key
The first two keys will be used as good and wrong key respectively in
the trust discovery test. But we will need a separate key for singing
stage1 image, so that the key does not interfere in the tests. Signing
the stage1 image is necessary since recently for stage1 images passed
to rkt via --stage1-path.