store: check the ACI file's UID instead of the diskv dir
To check if we can remove an image we were checking the UID of the
diskv directory where the image is located.

This is wrong because several images can end up in the same diskv
directory and be fetched by different users, so we would be checking
against the first user that imported an image there.