stage1: allow pid1 to be moved to init.scope
pid1 was moved to init.scope in the systemd hierarchy but not in the
other cgroup hierarchies. This was because systemd-nspawn does not
delegate the other cgroup hierarchies and mount them read-only. rkt
make parts of the other cgroup hierarchies read-write but this was not
extended to init.scope. This patch extends that logic to init.scope.