tests: new test TestFlyVolumeVar
To reproduce the explosive number of bind mounts manually, you can run
the following command several times:

> rkt --insecure-options=image run \
>     --volume host-var,kind=host,source=/var \
>     --mount volume=host-var,target=/host_var \
>     image.aci

It uses the volume /var because rkt stores the pods in a subdirectory of
/var by default.

The new test TestFlyVolumeVar implemented in this patch reproduces the
issue automatically. Instead of using the volume /var, it uses the
volume ctx.DataDir() because the test framework stores the pods in that
directory.

Reproduces https://github.com/coreos/rkt/issues/2553