stage1: allow stopping apps without shutting down the pod
Previously, the app's reaper was triggering a pod shutdown when it
executed `systemctl exit ${status}` because all the other apps' reapers
have a Conflict relationship with `exit.target`.

Instead, we now rely on the relationship between the apps' reapers and
`shutdown.service`. That is, when all reapers exit, `shutdown.service`
will shut down the pod.

For more details see
https://github.com/coreos/rkt/blob/master/Documentation/devel/architecture.md