stage1: reorder iomux service
This changes the start order of the iomux service so it starts before
the app service instead of after (in the diff, "before" and "after" are
swapped because this code refers to the app service).

If the app has a pre-start hook that fails, iomux will never start and
we won't get any logs of the pre-start hook so users will have a hard
time debugging. With this change, iomux will be started before the app
service starts, capturing all the relevant logs.